requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
jinja2>=3.1.6
python-telegram-bot>=22.0
loguru>=0.7.3
bcrypt>=4.1.0
python-multipart>=0.0.9
